Last Commit

+++ FrontEnd/src/secondary -components/singleuser.js
@@ -1,33 +1,40 @@
 import React from "react";
 import axios from "axios";
 import '../componentsCss/alluser.css'
+import { useNavigate } from 'react-router-dom';
 
 
 const Userligne = ({user, onDeleteUser }) => {
-   /* const del = () => {
-      axios.delete(`http://127.0.0.1:3000/user/deluser/${props.single._id}`)
-      .then(res => {
-          console.log(res);
-          console.log(res.data);
-          props.setUsers(prevUsers => prevUsers.filter(user => user._id !== props.single._id));
-      })
-      .catch(error => console.error(error));
-    }*/
+
+
+    const navigate = useNavigate();
+
+
     const handleDeleteClick = () => {
-      onDeleteUser(user._id);
+        const confirmed = window.confirm('Are you sure you want to delete this user?');
+        if (confirmed) {
+          onDeleteUser(user._id);
+        }
+   
     }
+    const handleDailySheetClick = () => {
+       
+        navigate(`/Dash/PDsheet/${user.MonthlySheetid}?User=${user._id}&Name= : ${user.FirstName} ${user.LastName} `);
+
+
+    };
     return (<div className="tableuser">
 
         <div className="table-wrapper">
             <table className="fl-table">
 
                 <tbody>
-
-                    <td >{user.FirstName}</td>
-                    <td>{user.LastName}</td>
+              
+                <td > <img src={require('../assets/UserIcon.png')} style={{ position: 'absolute', left: '1%', marginTop: '-12px' }} ></img>{user.FirstName}</td>
+                  
                     <td>{user.Email}</td>
-                    <td>{user.Phone}</td>
-                    <td> <img src={require('../assets/trash.png')} height={"15px"} className="trash" onClick={handleDeleteClick}></img> </td>
+                  <td onClick={handleDailySheetClick} style={{cursor:'pointer'}}>  {user.Lastadd ? user.Lastadd.substring(0, 10) : "Not Avaiable"}</td>
+                    <td> <img style={{cursor : 'pointer'}} src={require('../assets/deleteUser.png')}  className="trash" onClick={handleDeleteClick}></img> </td>
